+++ addons/african_south/subconfig_cup/script_component.hpp
@@ -1,0 +1,4 @@
+#define COMPONENT african_south_cup
+#define COMPONENT_BEAUTIFIED African South CUP
+#include "\x\tacu\addons\main\script_mod.hpp"
+#include "\x\tacu\addons\main\script_macros.hpp"
